--- deepliif/options/__init__.py
@@ -1,1 +1,0 @@
-"""This package options includes option modules: training options, test options, and basic options (used in both training and test)."""
