+++ deepliif/util/__init__.py
@@ -1,1 +1,94 @@
 """This package includes a miscellaneous collection of useful helper functions."""
+import os
+import collections
+
+import torch
+import numpy as np
+from PIL import Image, ImageOps
+
+from .visualizer import Visualizer
+
+# Postfixes not to consider for segmentation
+excluding_names = ['Hema', 'DAPI', 'DAPILap2', 'Ki67', 'Seg', 'Marked', 'SegRefined', 'SegOverlaid', 'Marker', 'Lap2']
+# Image extensions to consider
+image_extensions = ['.png', '.jpg', '.tif']
+
+
+def allowed_file(filename):
+    name, extension = os.path.splitext(filename)
+    image_type = name.split('_')[-1]  # Read image type
+
+    return extension in image_extensions and image_type not in excluding_names
+
+
+def chunker(iterable, size):
+    for i in range(size):
+        yield iterable[i::size]
+
+
+def tensor2im(input_image, imtype=np.uint8):
+    """"Converts a Tensor array into a numpy image array.
+
+    Parameters:
+        input_image (tensor) --  the input image tensor array
+        imtype (type)        --  the desired type of the converted numpy array
+    """
+    if not isinstance(input_image, np.ndarray):
+        if isinstance(input_image, torch.Tensor):  # get the data from a variable
+            image_tensor = input_image.data
+        else:
+            return input_image
+        image_numpy = image_tensor[0].cpu().float().numpy()  # convert it into a numpy array
+        if image_numpy.shape[0] == 1:  # grayscale to RGB
+            image_numpy = np.tile(image_numpy, (3, 1, 1))
+        image_numpy = (np.transpose(image_numpy, (1, 2, 0)) + 1) / 2.0 * 255.0  # post-processing: tranpose and scaling
+    else:  # if it is a numpy array, do nothing
+        image_numpy = input_image
+    return image_numpy.astype(imtype)
+
+
+def tensor_to_pil(t):
+    return Image.fromarray(tensor2im(t))
+
+
+Tile = collections.namedtuple('Tile', 'i, j, img')
+
+
+def output_size(img, tile_size):
+    return (max(round(img.width / tile_size) * tile_size, tile_size),
+            max(round(img.height / tile_size) * tile_size, tile_size))
+
+
+def generate_tiles(img, tile_size, overlap_size):
+    img = img.resize(output_size(img, tile_size))
+    # Adding borders with size of given overlap around the whole slide image
+    img = ImageOps.expand(img, border=overlap_size, fill='white')
+    rows = int(img.height / tile_size)  # Number of tiles in the row
+    cols = int(img.width / tile_size)  # Number of tiles in the column
+
+    # Generating the tiles
+    for i in range(cols):
+        for j in range(rows):
+            yield Tile(j, i, img.crop((
+                i * tile_size, j * tile_size,
+                i * tile_size + tile_size + 2 * overlap_size,
+                j * tile_size + tile_size + 2 * overlap_size
+            )))
+
+
+def stitch(tiles, tile_size, overlap_size):
+    rows = max(t.i for t in tiles) + 1
+    cols = max(t.j for t in tiles) + 1
+
+    width = tile_size * cols
+    height = tile_size * rows
+
+    new_im = Image.new('RGB', (width, height))
+
+    for t in tiles:
+        img = t.img.resize((tile_size + 2 * overlap_size, tile_size + 2 * overlap_size))
+        img = img.crop((overlap_size, overlap_size, overlap_size + tile_size, overlap_size + tile_size))
+
+        new_im.paste(img, (t.j * tile_size, t.i * tile_size))
+
+    return new_im
