+++ deepliif/util/util.py
@@ -20,31 +20,6 @@ def timeit(f):
     return wrap
 
 
-def tensor2im(input_image, imtype=np.uint8):
-    """"Converts a Tensor array into a numpy image array.
-
-    Parameters:
-        input_image (tensor) --  the input image tensor array
-        imtype (type)        --  the desired type of the converted numpy array
-    """
-    if not isinstance(input_image, np.ndarray):
-        if isinstance(input_image, torch.Tensor):  # get the data from a variable
-            image_tensor = input_image.data
-        else:
-            return input_image
-        image_numpy = image_tensor[0].cpu().float().numpy()  # convert it into a numpy array
-        if image_numpy.shape[0] == 1:  # grayscale to RGB
-            image_numpy = np.tile(image_numpy, (3, 1, 1))
-        image_numpy = (np.transpose(image_numpy, (1, 2, 0)) + 1) / 2.0 * 255.0  # post-processing: tranpose and scaling
-    else:  # if it is a numpy array, do nothing
-        image_numpy = input_image
-    return image_numpy.astype(imtype)
-
-
-def tensor_to_pil(t):
-    return Image.fromarray(tensor2im(t))
-
-
 def diagnose_network(net, name='network'):
     """Calculate and print the mean of average absolute(gradients)
 
@@ -121,6 +96,3 @@ def mkdir(path):
         os.makedirs(path)
 
 
-def chunker(iterable, size):
-    for i in range(size):
-        yield iterable[i::size]
